requirements.txt
Django==1.8
Markdown==2.6.2
django-filter==0.9.2
djangorestframework==3.1.1
wsgiref==0.1.2
